Exp-1/output/f00306_generate_text.py
from typing import *
from transformers import AutoModelForCausalLM

def generate_text(inputs, max_new_tokens=100, do_sample=True, top_k=50, top_p=0.95):
    """Generate text using the specified model.

    Args:
        inputs (str): The input text to start the generation from.
        max_new_tokens (int, optional): The maximum number of tokens to generate. Defaults to 100.
        do_sample (bool, optional): Whether to use sampling for generation. Defaults to True.
        top_k (int, optional): The number of highest probability tokens to keep for sampling. Defaults to 50.
        top_p (float, optional): The cumulative probability threshold for sampling. Defaults to 0.95.
    """
    model = AutoModelForCausalLM.from_pretrained("my_awesome_eli5_clm-model")
    outputs = model.generate(inputs, max_new_tokens=max_new_tokens, do_sample=do_sample, top_k=top_k, top_p=top_p)
