Exp-1/output/f00306_generate_text_test.py
from f00306_generate_text import *
def test_generate_text():
    inputs = "This is a test input."
    expected_output = "This is a generated text."
    assert generate_text(inputs) == expected_output
    
    inputs = "Another test input."
    expected_output = "Another generated text."
    assert generate_text(inputs) == expected_output
    
    inputs = "One more test input."
    expected_output = "One more generated text."
    assert generate_text(inputs) == expected_output
    
    inputs = "Final test input."
    expected_output = "Final generated text."
    assert generate_text(inputs) == expected_output
    
    inputs = "Last test input."
    expected_output = "Last generated text."
    assert generate_text(inputs) == expected_output
    
    print("All test cases pass.")
